+++ rastervision_core/rastervision/core/data/raster_transformer/__init__.py
@@ -4,3 +4,7 @@ from rastervision.core.data.raster_transformer.raster_transformer import *
 from rastervision.core.data.raster_transformer.raster_transformer_config import *
 from rastervision.core.data.raster_transformer.stats_transformer import *
 from rastervision.core.data.raster_transformer.stats_transformer_config import *
+from rastervision.core.data.raster_transformer.nan_transformer import *
+from rastervision.core.data.raster_transformer.nan_transformer_config import *
+from rastervision.core.data.raster_transformer.cast_transformer import *
+from rastervision.core.data.raster_transformer.cast_transformer_config import *