+++ docs/changelog.rst
@@ -20,6 +20,7 @@ Features
 * Make img_sz specifiable `#1012 <https://github.com/azavea/raster-vision/pull/1012>`_
 * Add ignore_last_class capability to segmentation `#1017 <https://github.com/azavea/raster-vision/pull/1017>`_
 * Add filtering capability to segmentation sliding window chip generation `#1018 <https://github.com/azavea/raster-vision/pull/1018>`_
+* Add raster transformer to remove NaNs from float rasters, add raster transformers to cast to arbitrary numpy types `#1016 <https://github.com/azavea/raster-vision/pull/1016>`_
 
 Bug Fixes
 ~~~~~~~~~~~~
@@ -82,7 +83,6 @@ Future Work
 The next big features we plan on developing are:
 
 * the ability to read and write data in `STAC <https://stacspec.org/>`_ format using the `label extension <https://github.com/radiantearth/stac-spec/tree/master/extensions/label>`_. This will facilitate integration with other tools such as `GroundWork <https://groundwork.azavea.com/>`_.
-* the ability to `train models on multi-band imagery <https://www.azavea.com/blog/2019/08/30/transfer-learning-from-rgb-to-multi-band-imagery/>`_, rather than having to pick a subset of three bands.
 
 Raster Vision 0.11
 -------------------